refactor stat.py
